scripts/policy/requirements.txt
et-xmlfile==1.0.1
google-api-python-client==1.5.3
httplib2==0.9.2
jdcal==1.2
oauth2client==3.0.0
openpyxl==2.3.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyDrive==1.2.1
python-dateutil==2.5.3
PyYAML==3.12
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
xlrd==1.0.0
